Test cases and implementation for subsequence were wrong.
Actually both test cases and the implementation of the `Subsequence.index_is_in` function were wrong. Firstly, the entry `start_index + length` is not part of the subsequence, because that it would include `length+1` elements. This has to be changed both in the implementation and the test.

Also, the second test case is wrong. A sequence with lenght 0 does not contain any elements of course. I change the test-cases and added a test for a lenght 1 subsequence.